hf_space/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
pytorch-lightning>=2.1.0
timm>=0.9.0
gradio>=4.0.0
Pillow>=10.0.0
loguru>=0.7.0
numpy>=1.24.0
opencv-python>=4.5.0
huggingface_hub>=0.20.0
